requirements-render.txt
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.10.0 ; python_version >= "3.13"
autogen-agentchat==0.7.4 ; python_version >= "3.13"
autogen-core==0.7.4 ; python_version >= "3.13"
autogen-ext==0.7.4 ; python_version >= "3.13"
backrefs==6.0.1 ; python_version >= "3.13"
bleach==6.2.0 ; python_version >= "3.13"
blinker==1.9.0 ; python_version >= "3.13"
cachetools==5.5.2 ; python_version >= "3.13"
certifi==2025.8.3 ; python_version >= "3.13"
cffi==1.17.1 ; python_version >= "3.13" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.3 ; python_version >= "3.13"
click==8.2.1 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
cryptography==45.0.6 ; python_version >= "3.13"
distro==1.9.0 ; python_version >= "3.13"
flask-cors==5.0.0 ; python_version >= "3.13"
flask==3.1.2 ; python_version >= "3.13"
google-auth==2.40.3 ; python_version >= "3.13"
google-genai==1.32.0 ; python_version >= "3.13"
h11==0.16.0 ; python_version >= "3.13"
httpcore==1.0.9 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
importlib-metadata==8.7.0 ; python_version >= "3.13"
itsdangerous==2.2.0 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
jiter==0.10.0 ; python_version >= "3.13"
jsonref==1.1.0 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
memory-profiler==0.61.0 ; python_version >= "3.13"
oauthlib==3.3.1 ; python_version >= "3.13"
openai==1.102.0 ; python_version >= "3.13"
opentelemetry-api==1.36.0 ; python_version >= "3.13"
pillow==11.3.0 ; python_version >= "3.13"
protobuf==5.29.5 ; python_version >= "3.13"
psutil==6.1.1 ; python_version >= "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pycparser==2.22 ; python_version >= "3.13" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.13"
pydantic==2.11.7 ; python_version >= "3.13"
pyjwt==2.10.1 ; python_version >= "3.13"
pyotp==2.9.0 ; python_version >= "3.13"
python-bitvavo-api==1.4.3 ; python_version >= "3.13"
python-dotenv==1.1.1 ; python_version >= "3.13"
redis==6.4.0 ; python_version >= "3.13"
regex==2025.8.29 ; python_version >= "3.13"
requests-oauthlib==2.0.0 ; python_version >= "3.13"
requests==2.32.5 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
setuptools==80.9.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
tenacity==9.1.2 ; python_version >= "3.13"
tiktoken==0.11.0 ; python_version >= "3.13"
tomli==2.2.1 ; python_version >= "3.13"
tqdm==4.67.1 ; python_version >= "3.13"
typing-extensions==4.15.0 ; python_version >= "3.13"
typing-inspection==0.4.1 ; python_version >= "3.13"
urllib3==2.5.0 ; python_version >= "3.13"
webencodings==0.5.1 ; python_version >= "3.13"
websocket-client==1.8.0 ; python_version >= "3.13"
websockets==15.0.1 ; python_version >= "3.13"
werkzeug==3.1.3 ; python_version >= "3.13"
zipp==3.23.0 ; python_version >= "3.13"
